separate load_ion_profile | fix

+++ code/dataio/beam.h
@@ -1,2 +1,2 @@
 void load_beam(double *XR, double *XZ, double *XT, double *VR, double *VZ, double *VT, beam_prop beam, shot_prop shot, run_prop run);
-void load_ion_profile(char* shotname, double *prof_size, double *prof_r, double *prof_d, double *profx_r, double *profx_d);
+void load_ion_profile(char* shotname, int *prof_size, double *prof_r, double *prof_d, double *profx_r, double *profx_d);
